Drop unused rebuildFromAuthoritativeSources, scope to diagnostic needs
The diagnostic only needs the cheap totals path; the chronological
transaction stream and its types were landed ahead of any caller. Strip
the unused export, helpers, types, and tests; rename the file to match
its remaining responsibility; drop the unused export of
findEmbeddedWasteBalances.
